BJS2-5570: Update Calculator task based on review comments.

+++ src/main/java/faang/school/godbless/sprint2/lambda/calculator/Calculator.java
@@ -3,5 +3,5 @@ package faang.school.godbless.sprint2.lambda.calculator;
 @FunctionalInterface
 public interface Calculator {
 
-    Integer calculate(Integer numberFirst, Integer numberSecond);
+    Long calculate(Long numberFirst, Long numberSecond);
 }
